requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.40
fredapi>=0.5.0
pandas_datareader>=0.10.0
pandera>=0.20.0
pyarrow>=15.0.0
python-dotenv>=1.0.0
